+++ smartsheet/models/__init__.py
@@ -122,3 +122,4 @@ from .widget import Widget
 from .widget_content import WidgetContent
 from .workspace import Workspace
 from .user_plan import UserPlan
+from .asset_shares_paginated_result import AssetSharesPaginatedResult