+++ tests/mock_api/test_mock_api_sights.py
@@ -1,7 +1,7 @@
 # pylint: disable=C0103,W0232
 
 import smartsheet
-from mock_api_test_helper import MockApiTestHelper, clean_api_error
+from tests.mock_api.mock_api_test_helper import MockApiTestHelper, clean_api_error
 
 
 class TestMockApiSheets(MockApiTestHelper):